test: configure dual Bun/Jest testing strategy
