+++ eslint.config.ts
@@ -6,6 +6,7 @@ import reactRecommended from "eslint-plugin-react/configs/recommended.js";
 import reactJsxRuntime from "eslint-plugin-react/configs/jsx-runtime.js";
 import reactHooks from "eslint-plugin-react-hooks";
 import nextPlugin from "@next/eslint-plugin-next";
+import jestPlugin from "eslint-plugin-jest";
 
 // const mdxPlugin = mdxNamespace.default; // Previous attempt
 // Now using mdxPlugin (the namespace import) directly
@@ -231,6 +232,25 @@ const config = tseslint.config(
       "@typescript-eslint/no-unsafe-argument": "off"
     }
   } as any, // Disable TS strict rules in test files
+
+  // Jest specific configuration
+  {
+    // Apply Jest plugin rules only to files matching the pattern
+    files: ["**/?(*.)+(jest.spec|jest.test).{js,jsx,ts,tsx}"],
+    plugins: {
+      jest: jestPlugin,
+    },
+    rules: {
+      // Use recommended Jest rules
+      ...(jestPlugin.configs.recommended as any).rules,
+      // Add any Jest-specific overrides here if needed
+    },
+    languageOptions: {
+      globals: {
+        ...globals.jest, // Add Jest globals
+      },
+    },
+  },
 );
 
 export default config; // Export the config array
