+++ package.json
@@ -25,9 +25,11 @@
     "validate": "bun run lint && bun run type-check",
     "clean": "rm -rf .next out node_modules",
     "cleanCache": "rm -rf .next/cache",
-    "test": "./scripts/run-tests.sh",
-    "test:watch": "./scripts/run-tests.sh --watch",
-    "coverage": "./scripts/run-tests.sh --coverage",
+    "test:bun": "bun test",
+    "test:jest": "./scripts/run-tests.sh --testMatch \"**/?(*.)+(jest.spec|jest.test).{js,jsx,ts,tsx}\"",
+    "test": "bun run test:bun && bun run test:jest",
+    "test:watch": "./scripts/run-tests.sh --watch --testMatch \"**/?(*.)+(jest.spec|jest.test).{js,jsx,ts,tsx}\"",
+    "coverage": "./scripts/run-tests.sh --coverage --testMatch \"**/?(*.)+(jest.spec|jest.test).{js,jsx,ts,tsx}\"",
     "test:smoke:blog": "./scripts/run-tests.sh __tests__/blog/blog.smoke.test.ts",
     "test:smoke:pages": "./scripts/run-tests.sh __tests__/app/pages.smoke.test.ts",
     "prepare": "node .husky/install.mjs"
@@ -74,7 +76,6 @@
     "@swc/core": "^1.10.1",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^14.2.1",
-    "@types/cheerio": "^1.0.0",
     "@types/jest": "^29.5.14",
     "@types/node": "^20.11.24",
     "@types/node-fetch": "^2.6.12",
@@ -87,11 +88,13 @@
     "bun-types": "^1.2.12",
     "eslint": "^9.26.0",
     "eslint-config-next": "^15.3.2",
+    "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-mdx": "^3.4.1",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
     "globals": "^16.1.0",
     "husky": "^9.1.7",
+    "jest": "^29.7.0",
     "jiti": "^2.4.2",
     "lint-staged": "^15.5.2",
     "node-fetch": "^2.7.0",
